Added DisplayMap and GetCityData components
